+++ src/image.rs
@@ -1,0 +1,1 @@
+pub struct Image {}
