+++ src/sync_structures.rs
@@ -1,0 +1,32 @@
+use ash::vk;
+
+pub struct SyncStructures {
+    pub present_semaphore: vk::Semaphore,
+    pub render_semaphore: vk::Semaphore,
+    pub render_fence: vk::Fence,
+}
+
+impl SyncStructures {
+    pub fn new(device: &ash::Device) -> Self {
+        unsafe {
+            let render_fence = device
+                .create_fence(
+                    &vk::FenceCreateInfo::builder().flags(vk::FenceCreateFlags::SIGNALED),
+                    None,
+                )
+                .unwrap();
+            let present_semaphore = device
+                .create_semaphore(&vk::SemaphoreCreateInfo::builder(), None)
+                .unwrap();
+            let render_semaphore = device
+                .create_semaphore(&vk::SemaphoreCreateInfo::builder(), None)
+                .unwrap();
+
+            Self {
+                present_semaphore,
+                render_semaphore,
+                render_fence,
+            }
+        }
+    }
+}
